+++ task.php
@@ -42,6 +42,17 @@ $reshook = $hookmanager->executeHooks('formObjectOptions', $parameters, $tmpEven
     <script>
 
         document.addEventListener('DOMContentLoaded', function () {
+            //Définition de la boite de dialog
+            var taskediteventmodal = $('#dialog-edit-event');
+
+			taskediteventmodal.dialog({
+                autoOpen: false,
+				autoResize:true,
+                close: function( event, ui ) {
+                   $('#calendar').fullCalendar('refetchEvents');
+                }
+            });
+            //fullcalendar
             var currentsource = '<?php echo dol_buildpath('/fullcalendar/script/interface.php', 1) ?>'+'?get=tasks';
             $('#calendar').fullCalendar({
                 header: {
@@ -91,12 +102,13 @@ $reshook = $hookmanager->executeHooks('formObjectOptions', $parameters, $tmpEven
                 , eventRender: function (event, element, view) {
                     var title = element.find('.fc-title').html();
                     element.find('.fc-title').html('<a class="url_title" href="'+event.url_title+'" onclick="event.stopPropagation();">'+title+'</a>');
+					element.find('.fc-content').prepend(event.headTask);
 
                     if ($().tipTip) // ou $.fn.tipTip, mais $.tipTip ne fonctionne pas
                     {
                         element.tipTip({
                             maxWidth: '600px', edgeOffset: 10, delay: 50, fadeIn: 50, fadeOut: 50
-                            , content: '<strong>'+event.title+'</strong><br />'+event.description
+                            , content: event.headTask+'<strong>'+event.title+'</strong><br />'+event.description
                         });
 
                         element.find('.classfortooltip').tipTip({maxWidth: '600px', edgeOffset: 10, delay: 50, fadeIn: 50, fadeOut: 50});
@@ -107,7 +119,7 @@ $reshook = $hookmanager->executeHooks('formObjectOptions', $parameters, $tmpEven
                             , show: {collision: 'flipfit', effect: 'toggle', delay: 50}
                             , hide: {delay: 50}
                             , position: {my: 'left+10 center', at: 'right center'}
-                            , content: '<strong>'+event.title+'</strong><br />'+event.description
+                            , content: event.headTask+'<strong>'+event.title+'</strong><br />'+event.description
                         });
 
                         // On remet les tooltips des liens désactivés par l'appel ci-dessus
@@ -140,6 +152,32 @@ $reshook = $hookmanager->executeHooks('formObjectOptions', $parameters, $tmpEven
                         $('#calendar').fullCalendar('refetchEvents');
                     });
                 }
+                ,
+				eventClick: function(info) {
+					$.ajax({
+                        url: '<?php echo dol_buildpath('/fullcalendar/script/interface.php', 1) ?>'
+                        , data: {
+                            get: 'task-popin'
+                            , fk_task: info.id
+                        }
+                    }).done(function (data) {
+                        $('#dialog-edit-event').html(data);
+                        taskediteventmodal.dialog('open');
+                        taskediteventmodal.dialog({
+                            height: 'auto', width: 'auto'
+                            , buttons: {
+                                '<?php echo $langs->trans('Update'); ?>': function () {
+                                    updateTask();
+                                    $(this).dialog('close');
+                                },
+                               '<?php echo $langs->trans('Cancel'); ?>': function () {
+                                    $(this).dialog('close');
+                                }
+                            }
+                        }); // resize to content
+                        taskediteventmodal.parent().css({'top': '20%'});
+                    });
+				}
                 , eventDrop: function (event, delta, revertFunc, jsEvent, ui, view) {
 
                     $.ajax({
@@ -154,7 +192,26 @@ $reshook = $hookmanager->executeHooks('formObjectOptions', $parameters, $tmpEven
                     });
                 }
             });
+
         });
+
+        // refresh event on modal close
+        $("#dialog-edit-event").on("hide.bs.modal", function (e) {
+             $('#calendar').fullCalendar('refetchEvents');
+        });
+
+        function updateTask() {
+            let data = $('#editableViewForm').serializeArray();
+            $.ajax({
+                url: '<?php echo dol_buildpath('/fullcalendar/script/interface.php', 1) ?>'
+                , data: {
+                    put: 'task-edit'
+                    , data: data
+                }
+            }).done(function () {
+                $('#calendar').fullCalendar('refetchEvents');
+            });
+        }
 
         function getFullCalendarHeight() {
             return $(window).height()-$('#id-right').offset().top-30;
@@ -176,6 +233,8 @@ $reshook = $hookmanager->executeHooks('formObjectOptions', $parameters, $tmpEven
     </script>
 <?php
 print '<div id="calendar"></div>';
+print '<div id="dialog-edit-event" title="'.$langs->trans('EditTask').'"></div>';
+
 
 llxFooter();